refactor(lowy): split agent-mock.ts into per-agent fixture modules
The Codex JSONL schema (changed twice in 944f19d, 431edd3) and
OpenCode SQLite schema evolve on independent axes. A unified
`agent-mock.ts` module complects them — touching Codex token
accounting shouldn't require opening OpenCode's SQL. Split into
`agent-mock-codex.ts`, `agent-mock-opencode.ts`, and a tiny
`agent-lifecycle.ts` carrying the one shared concept.
